backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiosqlite==0.19.0
pydantic==2.5.0
PyYAML==6.0.1
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2
